% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FocalRcppNew.R
\name{FocalRcppNew}
\alias{FocalRcppNew}
\title{Calculate heterogeneity matrix for an arbitrary raster layer}
\usage{
FocalRcppNew(
  x,
  NgbProb,
  NgbHet,
  fun,
  Dist,
  filename = "",
  na.rm = TRUE,
  padValue = NA,
  angle = "horizontal",
  parallelize = FALSE
)
}
\arguments{
\item{x}{is a raster layer}

\item{NgbProb}{is the probability neighbourhood - an integer smaller than the dimension of the raster
or NgbProb = "static"}

\item{NgbHet}{is the heterogeneity neighbourhood - an integer smaller than the dimension of the raster}

\item{fun}{The heterogeneity metric can be one of the following: "HomogeneityPosition",
"HomogeneitySpectral", "ContrastPosition", "ContrastSpectral", "DissimilarityPosition",
"DissimilaritySpectral", "Entropy", "WeightedEntropyAbsPosition",
"WeightedEntropyAbsSpectral", "WeightedEntropySqrPosition",
"WeightedEntropySqrSpectral", "RaoAbsSpectral", "RaoAbsPosition",
"RaoSqrSpectral", "RaoSqrPosition"}

\item{Dist}{is an integer}

\item{filename}{if the output raster should be written to a file}

\item{na.rm}{is logical and indicates if the NA values in the raster should be ignored}

\item{padValue}{defaults to NA}

\item{angle}{takes 5 options: "horizontal", "vertical", "diagonal45", "diagonal135", "all"}

\item{parallelize}{indicates if the calculation should be parallelized among cores}
}
\value{
A raster layer
}
\description{
This is the wrapper function that returns a raster layer with the chosen
heterogeneity metric
}
