% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patch}
\alias{patch}
\title{simulated patch data}
\format{
A matrix with 90 rows and 90 columns. Simulated patches.

\describe{
\item{\code{Simulated random patches}}{The background is random structure, depicting no spatial correlation, and structure is also random within patches.}
\item{\code{Value range}}{Data contains 20 gray levels.}
}
}
\usage{
patch
}
\description{
simulated patch data
}
\keyword{datasets}
