% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strucDivNest.R
\name{strucDivNest}
\alias{strucDivNest}
\title{Calculate spatial (horizontal) structural diversity for an arbitrary raster layer.}
\usage{
strucDivNest(
  x,
  wslI = NULL,
  wslO = NULL,
  dimB = FALSE,
  oLap = NULL,
  priorB = FALSE,
  domain = FALSE,
  dist = 1,
  angle = "all",
  rank = FALSE,
  fun,
  delta = 0,
  na.handling = na.pass,
  padValue = NA,
  aroundTheGlobe = FALSE,
  parallelize = FALSE,
  display_progress = FALSE,
  filename = "",
  ...
)
}
\arguments{
\item{x}{raster layer. Input raster layer for which 
horizontal structural diversity should be calculated.}

\item{wslI}{uneven integer. The window side length, 
wslI x wslI defines the size of the moving window.
The window must be smaller than the dimensions of the input raster. Default is NULL
in which case no nesting is used.}

\item{wslO}{an uneven integer smaller than the dimensions of the input raster 
and larger than wslI.}

\item{dimB}{is a vector of length 2 or a logical. This is the block size.
Default is FALSE in which case the image is not cut into blocks.}

\item{oLap}{an integer that defines the overlap between the blocks. It must
be at least 2*(wslO-1) or bigger. Blocks can overlap by a maximum of half the 
rows of blocks in row-direction, and by half the columns of blocks in column-direction.}

\item{priorB}{logical Should the blocks be used for prior estimates? 
If priorB = TRUE, then the spatial structure in a block serves as prior for 
spatial structural diversity estimates of the moving windows inside the block. 
If priorB = FALSE, then the blocks are only used to increase speed through 
parallelization, not as prior estimates.}

\item{domain}{logical. Should the spatial structure in the domain 
(i.e. the input raster) be used as prior estimate for the moving windows?}

\item{dist}{integer. The distance between two pixels that should be considered as a pair, 
defaults to dist = 1 (direct neighbors).}

\item{angle}{string. The angle on which pixels should be considered as pairs. 
Takes 5 options: "horizontal", "vertical", "diagonal45", "diagonal135", "all". 
"all" considers all of the 4 angles. Defaults to "all".}

\item{rank}{logical. Should pixel values be replaced with ranks in each GLCM? Defaults to FALSE.}

\item{fun}{function, the diversity metric. Takes one of the following: entropy,
entropyNorm, contrast, dissimilarity, or homogeneity.}

\item{delta}{integer, takes 3 options: 0, 1, or 2. 
Delta is the difference weight parameter, 
it defines how the differences between pixel values within a pixel pair should be weighted.  
If rank = TRUE, delta defines how the differences between ranks should be weighted.  
The default value is 0 (no weight). Set delta = 1 for absolute difference weight, 
or delta = 2 for squared difference weight. 
The delta parameter can only be set when the metric entropy is used. 
Dissimilarity automatically employs delta = 1, and contrast employs delta = 2.}

\item{na.handling}{na.omit or na.pass. 
If na.handling = na.omit, NAs are ignored, diversity metrics are calculated with less values. 
In this case the GLCM does not sum to 1.
If na.handling = na.pass and if there is at least one missing value inside the moving window,
an NA is assigned to the center pixel. Therefore, the diversity map will contain more 
NAs than the input raster layer.
Defaults to na.pass.}

\item{padValue}{numeric. The value of the padded cells at the edges of the input raster. 
Defaults to NA.}

\item{aroundTheGlobe}{logical. If the input raster goes around the whole globe, 
set aroundTheGlobe = TRUE, and the input raster will be "glued together" from both sides
to calculate diversity without edge effects on the sides.
Defaults to FALSE.}

\item{parallelize}{logical. Set to TRUE to parallelize the computation on multiple cores.
Default is FALSE. Please set to TRUE only for bigger problems.}

\item{display_progress}{logical}

\item{filename}{character. If the output raster should be written to a file, define file name (optional).}

\item{...}{possible further arguments.}
}
\value{
A raster layer
}
\description{
This is a wrapper function that returns a spatial structural diversity map 
as a raster layer. 
Pixel pairs are considered at user defined distances and angles within a moving window, 
the resulting frequency table is turned into a gray level co-occurrence matrix (GLCM).
Based on the GLCM, spatial structural diversity metrics are calculated, 
the metric value is assigned to the center pixel of the moving window. 
The output is a diversity map, which is returned as a raster layer.
}
\examples{

a <- raster::raster(matrix(rnorm(648), 18, 36))
raster::plot(a)
b <- StrucDiv(a, 3, fun = contrast, na.handling = na.omit, rank = FALSE)
raster::plot(b)

}
