% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_switch_function.R
\name{Diversity}
\alias{Diversity}
\alias{homogeneity}
\alias{dissimilarity}
\alias{contrast}
\alias{entropy}
\alias{entropyNorm}
\title{Diversity functions}
\usage{
homogeneity(
  rank,
  delta,
  Hetx,
  vMat_big = NULL,
  SpatMat,
  nrp,
  narm,
  display_progress = TRUE,
  ...
)

dissimilarity(
  rank,
  delta,
  Hetx,
  vMat_big = NULL,
  SpatMat,
  nrp,
  narm,
  display_progress,
  ...
)

contrast(
  rank,
  delta,
  Hetx,
  vMat_big = NULL,
  SpatMat,
  nrp,
  narm,
  display_progress,
  ...
)

entropy(
  rank,
  delta,
  Hetx,
  vMat_big = NULL,
  SpatMat,
  nrp,
  narm,
  display_progress,
  ...
)

entropyNorm(
  rank,
  delta,
  Hetx,
  vMat_big = NULL,
  SpatMat,
  nrp,
  narm,
  display_progress,
  ...
)
}
\arguments{
\item{rank}{logical. Should values be replaced with ranks in each co-occurrence 
matrix. Defaults to FALSE.}

\item{delta}{difference weight. The delta parameter can only be set when the metric entropy is used. 
The default value is 0. It can also take the values 1 for absolute value and
2 for squared. Contrast automatically employs delta = 2 and dissimilarity employs delta = 1.}

\item{Hetx}{is the diversity matrix that is returned by an internal function
to the \code{StrucDiv} function.}

\item{vMat_big}{is a matrix. Default is NULL.}

\item{SpatMat}{is the probability matrix that is returned by an internal function
to the \code{StrucDiv} function.}

\item{nrp}{integer. The number of possible pixel pairs. StrucDiv calculates it 
internally and passes it to the diversity functions.}

\item{narm}{logical. It is automatically set to 0 if na.handling = na.pass, and to 1 if na.handling = na.omit.}

\item{display_progress}{logical. Shows if the progress bar should be displayed.}

\item{...}{possible further arguments.}
}
\description{
The \code{homogeneity}, \code{contrast}, \code{dissimilarity} and \code{entropy} 
functions are used in the default configuration of \code{\link{StrucDiv}}.
}
\details{
This function is used internally and is called as an argument to the \code{StrucDiv}.
}
\examples{
\dontrun{
a <- raster::raster(matrix(rnorm(25), 5, 5))
hetx <- matrix(rnorm(25*9), 5*5, 3*3)
z <- matrix(runif(9, 0,1), 3, 3)
spatmat <- list(z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z, z)
}
}
