% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValuesWindow.R
\name{getValuesWindow}
\alias{getValuesWindow}
\title{Get the values from a raster object centered at each pixel}
\usage{
getValuesWindow(x, wsl, padValue, aroundTheGlobe, ...)
}
\arguments{
\item{x}{raster layer. Input raster layer for which}

\item{wsl}{integer. size of neighbourhood}

\item{padValue}{atomic. If a pixel is on the edges of an image, what value should it be padded with.}

\item{aroundTheGlobe}{logical. Does the image go around the globe.}
}
\value{
A matrix
}
\description{
Modified R Code from raster package - getValuesFocal,
to retrieve values of defined neighbourhood. Returns one row per pixel 
(row containing pixel neighbourhood (wsl)
}
